Aura overlay scrollbars shouldn't invalidate thumb on scroll or enabled state.
We still invalidate the scroll controls on ScrollableArea, which is used if the
scrollbars are painted into the ScrollableArea itself, rather than composited
layers.

The thumbNeedsRepaint property on Scrollbar is used to repaint a scrollbar thumb
that's actually changed appearance. This isn't actually needed when the thumb
simply changes scroll offset. Nor is it needed when the enabled/disabled state
changes since the compositor knows how to hide them itself. Thus, a composited
scrollbar now needs to repaint only when the hovered or pressed state changes
or if there's a geometry change.

BUG=669670

Review-Url: https://codereview.chromium.org/2590163004
Cr-Original-Commit-Position: refs/heads/master@{#440450}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 65f5be3dadd8a0a752effd7073a12f9ecdb1c472