Fix crashes when FileWriter methods are called after context is destroyed.
In r439956 FileWriter was changed to reset writer() when the context is
destroyed, but that meant that now calling any of its methods after the
context was destroyed would crash. This fixes that by making sure the
context is still alive when any of its javascript methods are called.

BUG=676407
BUG=656453

Review-Url: https://codereview.chromium.org/2601543002
Cr-Original-Commit-Position: refs/heads/master@{#440465}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2726ac3437f9524bc027021f672bb69a343cc69a