SharedWorker: Make shared workers keyed by a pair of url and name
This patch makes shared workers keyed by a pair of url and name, and deprecates
URLMismatchError that was deleted from the spec:
https://github.com/whatwg/html/pull/175

Before this change, shared workers throws URLMismatchError when a given
worker's name is equal to an existing shared worker but a given script URL is
different. After this change, a new shared worker is simply created in the case.

BUG=538914

Review-Url: https://codereview.chromium.org/2604733002
Cr-Original-Commit-Position: refs/heads/master@{#440729}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a2f777316c98a1bcc06f2d34120b7bec2c86a237