Always use html_diff to generate -pretty-diff.html files.
The idea of this patch is to start using html_diff for -pretty-diff.html
files; once these files are always available then the wdiff-related code
can be removed and there will still be some kind of HTML diff on Windows.

In order to keep this change minimal, the files are still called
-pretty-diff.html and "has_pretty_patch" is still set in the results JSON,
but after this, the "has_pretty_patch" and "has_wdiff" keys in the results
JSON should be removed.

BUG=672651

Review-Url: https://codereview.chromium.org/2580343003
Cr-Original-Commit-Position: refs/heads/master@{#440765}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 498957556586803f87827652555e1b29d7b4866a