Initialize member array field.
In RuleData we have array field, which is not initialized at
construction of the object. And after object constructed, we iterate
through this array by pointer. This code look dangerous and MSan detect
this problem.
To be safe during this iteration need initialize this array. It was
done.

R=skobes
BUG=None

Review-Url: https://codereview.chromium.org/2600123002
Cr-Original-Commit-Position: refs/heads/master@{#440838}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2d88572a7e87c280d1b1fdd222ebec771b8be482