Minor clean-up on WindowProxy::disposeContext()
We don't need to check isContextInitialized() in clearForClose() and
clearForNavigation() because it is already checked in WindowProxy::disposeContext().

Also this CL moves the ScriptState::Scope into disposeContext()
Though we don't need to enter ScriptState::Scope in case of clearForClose()
by accident, it looks safer to enter the scope always.

BUG=677253

Review-Url: https://codereview.chromium.org/2601033002
Cr-Original-Commit-Position: refs/heads/master@{#440839}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 488433ae435fbcd7ea4eac85f7042e7623f6bdef