Remove existingWindowProxy()
The only call site of existingWindowProxy() is V8Initializer::messageHandlerInMainThread().
However, as commented in the code, that code should not be needed because
if the context is during initialization it should already have bailed out at
the if(!scriptState->contextIsValid()) check at the very beginning of
V8Initializer::messageHandlerInMainThread().

If the reasoning is correct, we can simply remove all code around existingWindowProxy().

BUG=677253

Review-Url: https://codereview.chromium.org/2607743002
Cr-Original-Commit-Position: refs/heads/master@{#440840}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 45830933ad06e7f8a213e97269522469b1602863