Load data URI images in an async way according to spec
According to
https://github.com/ResponsiveImagesCG/picture-element/issues/223#issuecomment-69772765
we seem to be the only browser that loads data URIs immediately (as if
they were cached).
It was introduced as part of the move to async image loading:
https://codereview.chromium.org/200923002

This is now causing issues:
https://code.google.com/p/chromium/issues/detail?id=469131

This CL changes this behavior to be spec compliant and load data URIs
asynchronously, like other resources.

This is relanding https://codereview.chromium.org/1492303003 and
https://codereview.chromium.org/2173003002/

BUG=514206

Review-Url: https://codereview.chromium.org/2592113003
Cr-Original-Commit-Position: refs/heads/master@{#441184}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6403174c6111ea0947dc7cf8b1eb8a6f9e8546c