Refactor and add comments to webkitpy/w3c/update_w3c_test_expectations.py.
Background: This is the module that's responsible for updating baselines and test expectations when auto-importing. The basic process is:

 1. Collate test failure information.
 2. Where possible, download text baselines.
 3. For everything else, add lines to TestExpectations.

This is a refactoring CL and shouldn't affect behavior.

The purpose of this CL here is to prepare for another change involving this file (for http://crbug.com/676196).

Review-Url: https://codereview.chromium.org/2601083002
Cr-Original-Commit-Position: refs/heads/master@{#441187}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 27cb5fbb2d1846856ec1dc2ac0095bee8232ee03