Add example cases in unit tests for tests with "Slow" expectations.
This CL just expands the unit tests for:
  layout_test_results
  rebaseline_cl
  update_w3c_test_expectations:

The purpose of this was to investigate http://crbug.com/676196 and see whether slow tests were not rebaselined as expected due to being treated differently than results with expected: "Pass" and actual: "Text".

The result was that it appear that slow results should be handled as expected.

This CL doesn't change behavior -- it just adds more test methods to unit tests.

BUG=676196

Review-Url: https://codereview.chromium.org/2605103002
Cr-Original-Commit-Position: refs/heads/master@{#441191}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b341f101e1123f2eb9c67c074d8c1a5036fc46b2