Use atomics when accessing m_didSetLooping
Add setter and getter methods for accessing m_didSetLooping so that we
can access it atomically. The audio thread was reading this while the
main thread could be writing it.

Manually verified the test case from the bug that this fix works. (Need to reload the page a few times to trigger the original bug.)

BUG=676874
TEST=

Review-Url: https://codereview.chromium.org/2615463002
Cr-Original-Commit-Position: refs/heads/master@{#441304}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 310673a6eca24e30371c29593b4385fd3983ad1d