PlzNavigate: Pass the encoded resource transfer size to the renderer.
This should fix the http/tests/misc/resource-timing-sizes-tags.html layout test.
The other test http/tests/inspector/resource-har-conversion.html needs an expectation
change to allow the custom header which is passed for PlzNavigate. Will do that in a later
patchset.

Changes on this patch are to implement the GetTotalReceivedBytes() override for the StreamURLRequestJob class

BUG=673745

Review-Url: https://codereview.chromium.org/2612683002
Cr-Original-Commit-Position: refs/heads/master@{#441409}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a9167f903f8aae02ca586fb920348a71d573544