Rename HTMLElement::clickForBindings to click.
There's no reason to use a custom name for this, it's the only method
with this name and is implementing the spec'ed behavior. The name was
introduced in https://codereview.chromium.org/1285793004 when the
behavior was fixed, but there was no reason to change the name then.

Review-Url: https://codereview.chromium.org/2617593002
Cr-Original-Commit-Position: refs/heads/master@{#441542}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eebe629b0ff42cb0b6f787f0c17c858f0fc25cb2