Skip importing domxpath/xml_xpath_runner.html due to large file.
Currently, the large (2.2MB) file domxpath/xml_xpath_tests.xml
hasn't been imported because the auto-importer can't handle it
because Rietveld can't handle very large text files.

The auto-importer marked this test as fail, but instead we should
skip trying to import it for now since the reason why it can't
be run is not a problem with the test itself, but a problem with
the importer.

BUG=676491

Review-Url: https://codereview.chromium.org/2618563002
Cr-Original-Commit-Position: refs/heads/master@{#441550}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7c8e22a0a0e02013b973443aada2975d7d915e40