Do not skip ink for ideographic scripts
This patch changes not to skip ink for ideographic scripts, where
skipping ink is inappropriate.

In order to know when not to apply ink skipping, this patch adds an
optional vector to GlyphBuffer containing the glyph indexes that are
exempt from ink skipping.

BUG=677093

Review-Url: https://codereview.chromium.org/2598393002
Cr-Original-Commit-Position: refs/heads/master@{#441593}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8584def66acd7836e089a59a598f9f44e50ace7b