Remove ScriptController::namedItemAdded/Removed
They are unnecessary indirection to WindowProxy::namedItemAdded/Removed.

Also this CL updates a dcheck in WindowProxy::namedItemAdded/Removed to verify
that those methods are called after the window proxy is initialized.

(The check will be replaced with a better lifecycle model in a follow-up CL.)

BUG=677253

Review-Url: https://codereview.chromium.org/2616903002
Cr-Original-Commit-Position: refs/heads/master@{#441803}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5807dc6e5f6be9c395dc81b43335055f2d6a2a97