Migrate WTF::Vector::append() to ::push_back() [part 15 of N]
As discussed on platform-architecture-dev [0], there is interest in
normalizing the methods of WTF classes to better align with std classes.
This CL replaces append() with push_back() in Source/web/.

There are too many references to change them all at once, so this CL is
part 15 of many.

There are no functional changes.

[0] https://groups.google.com/a/chromium.org/d/topic/platform-architecture-dev/I7jnz4p1h84/discussion

BUG=662431

Review-Url: https://codereview.chromium.org/2612903007
Cr-Original-Commit-Position: refs/heads/master@{#441830}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c41acedb493c6d405f2f9bb665e6d366e33a45dc