Insert a line break when space is inserted at the top of wrapped lines
This patch changes to insert a line break when user inserts a space at
the top of wrapped lines, so that the inserted space appear at the
caret position.

The change is for plaintext only for now, so that we can collect
feedback and discuss with editor developers for what to do with
contenteditable.

BUG=76113

Review-Url: https://codereview.chromium.org/2618613004
Cr-Original-Commit-Position: refs/heads/master@{#441848}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3b600981286b135632371477f902214c55a1724