Fix memory leak with ScriptedIdleTaskController
If there is an untriggered idle task when the document is gone, it
introduces memory leak because IdleRequestCallbackWrapper keeps a
reference to the ScriptedIdleTaskController, making it live longer
than the document.

This patch fixes the issue by letting IdleRequestCallbackWrapper
store a WeakPersistent instead.

BUG=595155

Review-Url: https://codereview.chromium.org/2618893002
Cr-Original-Commit-Position: refs/heads/master@{#441929}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 114c62b99fd57ac1cf7c0da1150608fd5f68d3bf