Convert budget LayoutTests to use promise_test.
The infrastructure of async_test LayoutTests was changed recently, and
one effect is that an assert inside an event listener will cause the
test to time out instead of returning the error. This makes the test
much more difficult to debug failures. Additionally, promise_test is just
nicer when it's easy to return a promise.

This CL updates the existing budget tests which run in service workers to
use promise_test. It also cleans up a bit of style.

BUG=617971

Review-Url: https://codereview.chromium.org/2600733002
Cr-Original-Commit-Position: refs/heads/master@{#441931}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 66cd34190886e221b7d79beac0b19cf6b2ebf633