Stop hit testing bounds of an InlineFlowBox that overflows its containing box.
Previously, hit testing for a (root) InlineFlowBox would first test each of its
component boxes, then fall back to testing the bounds of the InlineFlowBox
against the hit test point. If the bounds of the InlineFlowBox contained
the hit test point, the hit test would succeed for that InlineFlowBox.

This is bad in the situation when an InlineFlowBox exceeds the bounds of its
containing LayoutBoxFlow, yet none of the component line boxes intersect the
hit test point.

Fix this by passing the bounds of the LayoutBoxFlow to clip the rect testing
the bounds of the InlineFlowBox.

BUG=417667

Review-Url: https://codereview.chromium.org/2605343003
Cr-Original-Commit-Position: refs/heads/master@{#442144}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b1c8f02315b33db1e6a6a28babcea4864d2e5b0