Fix displayed duration after onDurationChange
MediaControls::onDurationChange updates m_timeline, but does not update
the displayed duration. This means that HLS streams (and propably other
types of media) always have a duration of '0:00', because we don't know
the duration until we have started playing the media (and we initially
report a 0s duration).

This change fixes onDurationChange.

BUG=679049
TEST= Manually confirmed using a test HLS stream

Review-Url: https://codereview.chromium.org/2617993003
Cr-Original-Commit-Position: refs/heads/master@{#442330}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e9a62259f89620d6c31829937af7dd59b486f10