IDBObjectStore/IDBIndex.getAll/getAllKeys should have [EnforceRange]
Per spec, the |count| argument for the getAll() and getAllKeys()
method should have the [EnforceRange] extended attribute which makes
the method throw if an invalid number or number out of range is passed
(e.g. not a number, infinity, negative, etc) rather than silently
coercing as is the default (for historical reasons).

Blink was just missing these attributes on three of the four methods.
Firefox has them. Added the attributes and test cases.

BUG=679465
R=pwnall@chromium.org

Review-Url: https://codereview.chromium.org/2627453002
Cr-Original-Commit-Position: refs/heads/master@{#442419}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 99435bd5b9707422116b5aa06e4fc81adbeb34b7