Use WebURLResponse to pass the navigation preload response from SWContextClient to FetchEvent.
Currently we use WebServiceWorkerResponse to pass the navigation preload
response from ServiceWorkerContextClient::NavigationPreloadRequest to FetchEvent.

But to show the network timing information in DevTool's network tab, we need
the timing information in ResourceResponseInfo in blnik side. So this cl change
the code to use WebURLResponse which contains the timing information.

BUG=649558

Review-Url: https://codereview.chromium.org/2618313003
Cr-Original-Commit-Position: refs/heads/master@{#442813}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b9937a428eaf8d98b0ad9e05707ab2d0559a62b3